requirements.txt
streamlit==1.41.1
httpx==0.28.1
python-dotenv==1.0.1
google-genai==1.5.0
